+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 
 - Debugging mode now highlights the line that is to be run rather rather than the line pointed by program counter.
 - Fix Jump instruction bug that made the whole UI stop if it exists on first line 
+- Assembled Tab will automatically show assembled code once program is loaded. No need to manually press assemble code button which has also fixed alot of bugs related to assemble mode.